% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crequire.r
\name{crequire}
\alias{crequire}
\title{A custum require function}
\usage{
crequire(package, character.only = TRUE,
  repos = "http://cran.us.r-project.org")
}
\arguments{
\item{package}{Package name, is passed to require. No default}

\item{character.only}{Defaults to TRUE. Passed to require}

\item{repo}{Defaults to http://cran.us.r-project.org}
}
\description{
This function is a wrapper of the base require function that first tests if the desired package is installed, if not, installs it, and then requires it.
}
\keyword{require}
